Subtract a safety margin from the search count
If the search returns fewer results than we asked for, we assume that no more can be found. However, since CQP may actually return fewer results than we asked for (even though more can actually be found), we subtract a margin before deciding whether to continue searching to be on the safe side.