Check the last, not first, step for number of cpus
This is because the first step only searches the first 5 mill. words, and in corpora with very large texts there may be fewer cpu-assigned blocks in this step than there are cpus. Since the last step searches the entire remainder of the corpus, using the last step means a much lower risk of recalculating the bounds each time.